pyproject.toml
[project]
name = "tableau-mcp-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.55",
    "httpx>=0.28.1",
    "mcp>=1.18.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
